requirements.txt
flask>=3.0
flask-cors>=4.0
python-dotenv>=1.0
mistralai>=1.0
pdf2image>=1.17
Pillow>=10.0
gunicorn>=21.2.0
